requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-decouple==3.8
whitenoise==6.12.0
gunicorn==25.3.0
Pillow==11.2.1
